Break app-data into multiple files
